+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/content/CreativeModeTabItems.java
@@ -50,8 +50,6 @@ public final class CreativeModeTabItems {
             Blocks.INSTANCE.getFluidStorageBlock(variant)
         ));
         itemConsumer.accept(Blocks.INSTANCE.getMachineCasing());
-        itemConsumer.accept(Blocks.INSTANCE.getQuartzEnrichedIronBlock());
-        itemConsumer.accept(Blocks.INSTANCE.getQuartzEnrichedCopperBlock());
         itemConsumer.accept(Blocks.INSTANCE.getStorageMonitor());
         Items.INSTANCE.getNetworkTransmitters().stream().map(Supplier::get).forEach(itemConsumer);
         Items.INSTANCE.getNetworkReceivers().stream().map(Supplier::get).forEach(itemConsumer);